requirements.txt
python-dotenv
pydantic
requests

apache-airflow

fastapi
uvicorn

pymongo
mysql-connector-python

langchain
langchain-community
langchain-openai

openai

konlpy